backend/requirements.txt
Django>=6.0
Django REST Framework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.3
Pillow>=10.0
reportlab>=4.0
django-cors-headers>=4.3
